Fix warnings - unnecessary casts and utilizing values method on iterator.

+++ crates/readstat/src/rs_data.rs
@@ -340,7 +340,7 @@ impl ReadStatData {
             .set_row_offset(Some(self.chunk_row_start.try_into().unwrap()))?
             .parse_sas7bdat(ppath, ctx);
 
-        match FromPrimitive::from_i32(error as i32) {
+        match FromPrimitive::from_i32(error) {
             Some(ReadStatError::READSTAT_OK) => Ok(()),
             Some(e) => Err(From::from(format!(
                 "Error when attempting to parse sas7bdat: {:#?}",
